refactor: 각 컴포넌트 상단에 정의되어 있던 interface type.ts 파일에서 정의 후 import 해오도록  수정
