+++ client/src/components/Select/SelectContent.tsx
@@ -3,14 +3,11 @@ import type { HTMLAttributes, ReactNode, Ref } from 'react';
 import React, { forwardRef, useMemo } from 'react';
 import { useSelectContext } from './SelectContext';
 import { s_selectContent } from './Select.styles';
+import type { SelectContentProps } from './type';
 
 /**
  * SelectContent 컴포넌트
  */
-interface SelectContentProps extends HTMLAttributes<HTMLDivElement> {
-    children: ReactNode;
-    sx?: CSSObject;
-}
 
 function SelectContent(
     { children, sx, ...props }: SelectContentProps,
