+++ client/src/components/Select/SelectGroup.tsx
@@ -2,14 +2,11 @@ import type { CSSObject } from '@emotion/react';
 import type { HTMLAttributes, ReactNode } from 'react';
 import React from 'react';
 import { s_selectGroup } from './Select.styles';
+import type { SelectGroupProps } from './type';
 
 /**
  * SelectGroup 컴포넌트
  */
-interface SelectGroupProps extends HTMLAttributes<HTMLDivElement> {
-    children: ReactNode;
-    sx?: CSSObject;
-}
 
 function SelectGroup({ children, ...props }: SelectGroupProps) {
     return (
