+++ client/src/components/Select/SelectLabel.tsx
@@ -2,14 +2,11 @@ import type { CSSObject } from '@emotion/react';
 import type { HTMLAttributes, ReactNode } from 'react';
 import React from 'react';
 import { s_selectLabel } from './Select.styles';
+import type { SelectLabelProps } from './type';
 
 /**
  * SelectLabel 컴포넌트
  */
-interface SelectLabelProps extends HTMLAttributes<HTMLLabelElement> {
-    children: ReactNode;
-    sx?: CSSObject;
-}
 
 function SelectLabel({ children, ...props }: SelectLabelProps) {
     return (
